Refactor duplicated functions
Centralize duplicate yaml_include and flatten_navigation functions from
build_multiversion.py and conf.py into a shared build_utils.py module.
This eliminates redundant code and completely removes import name
collision risks with third-party packages. Also adds detailed inline
documentation for the page file resolution and dynamic placeholder
generation block.

Generated-by: Gemini 3.0 Pro
